requirements.txt
flask
pymupdf
fpdf
python-dotenv
google-genai